+++ flags/tests/test_forms.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from flags.conditions import CONDITIONS, register
+from flags.conditions.registry import _conditions, register
 from flags.forms import FlagStateForm
 
 
@@ -27,13 +27,23 @@ class FormTestCase(TestCase):
             },
         )
 
+    def test_bad_data(self):
+        form = FlagStateForm(
+            {"name": "FLAG_ENABLED", "condition": "boolean", "value": "flase"}
+        )
+        self.assertFalse(form.is_valid())
+        self.assertEqual(
+            form.errors,
+            {"value": ["Enter one of 'on', 'off', 'true', 'false', etc."]},
+        )
+
     def test_condition_choices_are_bound_late(self):
         @register("fake_condition")
         def fake_condition():
             return True  # pragma: no cover
 
         def cleanup_condition(condition_name):
-            del CONDITIONS[condition_name]
+            del _conditions[condition_name]
 
         self.addCleanup(cleanup_condition, "fake_condition")